버그 수정: 게시글 소유자만 게시글 수정 및 삭제 가능, 기능 추가: 뒤로가기

+++ myblog/post/views.py
@@ -42,6 +42,7 @@ def post_new(request):
             messages.error(request, '글 등록에 실패하였습니다.')
     else:
         form = PostForm()
+
     return render(request, 'post/post_form.html', {
         'form': form
     })
@@ -50,15 +51,26 @@ def post_new(request):
 @login_required
 def post_delete(request, id):
     post = get_object_or_404(Post, id=id)
+
+    if request.user != post.user:
+        messages.warning(request,'잘못된 접근입니다.')
+        return redirect('post:post_list')
+
     if request.method == "POST":
         post.delete()
         messages.success(request, '글 삭제를 완료했습니다.')
+
     return redirect('post:post_list')
 
 
 @login_required
 def post_edit(request, id):
     post = get_object_or_404(Post, id=id)
+
+    if request.user != post.user:
+        messages.warning(request,'잘못된 접근입니다.')
+        return redirect('post:post_list')
+
     if request.method == "POST":
         form = PostForm(request.POST, request.FILES, instance=post)
         if form.is_valid():
@@ -69,6 +81,7 @@ def post_edit(request, id):
             messages.error(request, '글 수정에 실패했습니다.')
     else:
         form = PostForm(instance=post)
+
     return render(request, 'post/post_form.html', {
         'form': form
     })
